+++ lib/readline.js
@@ -48,8 +48,6 @@ const {
   kClearScreenDown
 } = CSI;
 
-const { now } = process.binding('timer_wrap').Timer;
-
 const kHistorySize = 30;
 const kMincrlfDelay = 100;
 // \r\n, \n, or \r followed by something other than \n
@@ -409,9 +407,10 @@ Interface.prototype._normalWrite = function(b) {
   if (b === undefined) {
     return;
   }
+  const now = Date.now();
   var string = this._decoder.write(b);
   if (this._sawReturnAt &&
-      now() - this._sawReturnAt <= this.crlfDelay) {
+      now - this._sawReturnAt <= this.crlfDelay) {
     string = string.replace(/^\n/, '');
     this._sawReturnAt = 0;
   }
@@ -424,7 +423,7 @@ Interface.prototype._normalWrite = function(b) {
     this._line_buffer = null;
   }
   if (newPartContainsEnding) {
-    this._sawReturnAt = string.endsWith('\r') ? now() : 0;
+    this._sawReturnAt = string.endsWith('\r') ? now : 0;
 
     // got one or more newlines; process into "line" events
     var lines = string.split(lineEnding);
@@ -911,20 +910,22 @@ Interface.prototype._ttyWrite = function(s, key) {
   } else {
     /* No modifier keys used */
 
+    const now = Date.now();
+
     // \r bookkeeping is only relevant if a \n comes right after.
     if (this._sawReturnAt && key.name !== 'enter')
       this._sawReturnAt = 0;
 
     switch (key.name) {
       case 'return':  // carriage return, i.e. \r
-        this._sawReturnAt = now();
+        this._sawReturnAt = now;
         this._line();
         break;
 
       case 'enter':
         // When key interval > crlfDelay
         if (this._sawReturnAt === 0 ||
-            now() - this._sawReturnAt > this.crlfDelay) {
+            now - this._sawReturnAt > this.crlfDelay) {
           this._line();
         }
         this._sawReturnAt = 0;
